fetch/metadata/generated/speculation-rules-header.https.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/speculation-rules-header.sub.html
-->
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/fetch/metadata/resources/helper.sub.js"></script>
<body>
<script>
'use strict';

async function induceRequest(url, test) {
  const win = window.open();
  test.add_cleanup(() => win.close());
  win.location = `/common/blank.html?pipe=header(Speculation-Rules,"${encodeURIComponent(url)}")`;
}

const responseParams = {
  mime: 'application/speculationrules+json',
  body: '{}',
};

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
      });
}, 'sec-fetch-site - Same origin');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsCrossSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
}, 'sec-fetch-site - Cross-site');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsSameSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
}, 'sec-fetch-site - Same site');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite', 'httpsOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
}, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsOrigin', 'httpsSameSite', 'httpsOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
}, 'sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsCrossSite', 'httpsOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
}, 'sec-fetch-site - Cross-Site -> Same Origin');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsCrossSite', 'httpsSameSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
}, 'sec-fetch-site - Cross-Site -> Same-Site');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsCrossSite', 'httpsCrossSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
}, 'sec-fetch-site - Cross-Site -> Cross-Site');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsOrigin', 'httpsOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
      });
}, 'sec-fetch-site - Same-Origin -> Same Origin');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsOrigin', 'httpsSameSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
}, 'sec-fetch-site - Same-Origin -> Same-Site');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
}, 'sec-fetch-site - Same-Origin -> Cross-Site');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsSameSite', 'httpsOrigin'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
}, 'sec-fetch-site - Same-Site -> Same Origin');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsSameSite', 'httpsSameSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
}, 'sec-fetch-site - Same-Site -> Same-Site');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, ['httpsSameSite', 'httpsCrossSite'], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-site');
        assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
}, 'sec-fetch-site - Same-Site -> Cross-Site');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, [], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-mode');
        assert_array_equals(headers['sec-fetch-mode'], ['cors']);
      });
}, 'sec-fetch-mode');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, [], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_own_property(headers, 'sec-fetch-dest');
        assert_array_equals(headers['sec-fetch-dest'], ['speculationrules']);
      });
}, 'sec-fetch-dest');

promise_test((t) => {
  const key = '{{uuid()}}';

  return induceRequest(
      makeRequestURL(key, [], responseParams), t
    )
    .then(() => retrieve(key, {poll: true}))
    .then((headers) => {
        assert_not_own_property(headers, 'sec-fetch-user');
      });
}, 'sec-fetch-user');
</script>
